Split issues reducer in two
