+++ app/reducers/index.js
@@ -1,7 +1,8 @@
 import { combineReducers } from 'redux';
 import usersReducer from './user.reducer';
 import settingsReducer from './settings.reducer';
-import issuesReducer from './issues.reducer';
+import allIssuesReducer from './issues.reducer';
+import selectedIssueReducer from './issue.selected.reducer';
 import trackingReducer from './tracking.reducer';
 import projectReducer from './project.reducer';
 import timeEntryReducer from './timeEntry.reducer';
@@ -11,7 +12,10 @@ import { USER_LOGOUT } from '../actions/user.actions';
 const appReducer = combineReducers({
   user: usersReducer,
   settings: settingsReducer,
-  issues: issuesReducer,
+  issues: combineReducers({
+    all: allIssuesReducer,
+    selected: selectedIssueReducer
+  }),
   projects: projectReducer,
   tracking: trackingReducer,
   timeEntry: timeEntryReducer
